Only moderately related to menu sorting.
Added a window for displaying the sorting options. The footer was starting to get
a bit crowded.

+++ src/thread/io_thread.c
@@ -38,7 +38,7 @@ void run_io(
     {
     	// This check prevents lag between the read and display of stats
     	// without it the points on the graph can be several seconds behind.
-    	int minimumMet = cpuQueue->size < MIN_QUEUE_SIZE || memQueue->size < MIN_QUEUE_SIZE;
+    	u8 minimumMet = cpuQueue->size < MIN_QUEUE_SIZE || memQueue->size < MIN_QUEUE_SIZE;
     
     	if (minimumMet) 
     	{
